pyproject.toml
[tool.poetry]
name = "manager"
version = "0.1.0"
description = ""
authors = ["yalr098 <yvesromero1998@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.86.0"
uvicorn = "^0.19.0"
websockets = "^10.4"
pymongo = "^4.3.3"
motor = "^3.1.1"
python-dotenv = "^0.21.0"
APScheduler = "^3.9.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
